+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+* Add `_recovery_point` system space introduced in upcoming Tarantool 3.8
+  to crud.schema system spaces.
+
 ## [1.7.4] - 12-02-26
 
 ### Fixed