+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   to crud.schema system spaces.
 * Add `crud.locate()` method to find whether a tuple is in `memtx` or `vinyl` engine.
   Works for spaces managed by the enterprise `cooler` module.
+* Support for `mode`, `balance`, `prefer_replica` and `request_timeout` options in `crud.len`.
 
 ### Fixed
 * Allow read-only operations (`get`, `select`, `pairs`, `count`, `min`, `max`)